requirements.txt

Django==3.2
django-cors-headers==3.13.0
#django-environ==0.9.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
idna==3.3
# psycopg2-binary==2.9.3
#PyJWT==1.7.1
#python-dotenv==0.20.0
setuptools-rust==1.4.1
